# readtags: optimize the way to read fields