I measured performance of readtags by giving large tags file and found
strcmp in parseExtensionFields is a bottle neck.

This change reduces the number of calling strcmp in the function.

Without this change:
[yamato@slave]~/var/ctags-github% time readtags -t ~/.spelunker.d/kernel76/kernel76.tags  -l   > /dev/null (5 times)
<gs -t ~/.spelunker.d/kernel76/kernel76.tags  -l   > /dev/null
readtags -t ~/.spelunker.d/kernel76/kernel76.tags -l > /dev/null  3.62s user 0.29s system 99% cpu 3.918 total
readtags -t ~/.spelunker.d/kernel76/kernel76.tags -l > /dev/null  3.62s user 0.30s system 99% cpu 3.924 total
readtags -t ~/.spelunker.d/kernel76/kernel76.tags -l > /dev/null  3.64s user 0.28s system 99% cpu 3.930 total
readtags -t ~/.spelunker.d/kernel76/kernel76.tags -l > /dev/null  3.60s user 0.27s system 99% cpu 3.874 total
readtags -t ~/.spelunker.d/kernel76/kernel76.tags -l > /dev/null  3.65s user 0.25s system 99% cpu 3.909 total

With this change:
[yamato@slave]~/var/ctags-github% time readtags -t ~/.spelunker.d/kernel76/kernel76.tags  -l   > /dev/null (5 times)
readtags -t ~/.spelunker.d/kernel76/kernel76.tags -l > /dev/null  2.29s user 0.27s system 99% cpu 2.574 total
readtags -t ~/.spelunker.d/kernel76/kernel76.tags -l > /dev/null  2.24s user 0.27s system 99% cpu 2.518 total
readtags -t ~/.spelunker.d/kernel76/kernel76.tags -l > /dev/null  2.34s user 0.26s system 99% cpu 2.605 total
readtags -t ~/.spelunker.d/kernel76/kernel76.tags -l > /dev/null  2.27s user 0.26s system 99% cpu 2.530 total
readtags -t ~/.spelunker.d/kernel76/kernel76.tags -l > /dev/null  2.26s user 0.27s system 99% cpu 2.537 total

ctags with the change is about 1/3 faster than before.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>