@masatake 2019-02-26
> I'd rather go halfway myself, keep the initial computation for the length

I see. My understanding is that you are talking about:
```
if (colon - key == 4
```
.